pyproject.toml
[tool.poetry]
name = "route-distances"
version = "1.0.0"
description = "Models for calculating distances between synthesis routes"
authors = ["Genheden, Samuel <samuel.genheden@astrazeneca.com>"]
license = "MIT"

[tool.poetry.scripts]
train_lstm_model = "route_distances.tools.train_lstm_model:main"
prepare_aizynthfinder_output = "route_distances.tools.prepare_aizynthfinder_output:main"
cluster_aizynth_output = "route_distances.tools.cluster_aizynth_output:main"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
torch = "^1.7.1"
pytorch-tree-lstm = "^0.1.3"
pytorch-lightning = "^1.6.0"
optuna = "^2.5.0"
tables = "^3.6.0"
apted = "^1.0.3"
scikit-learn = ">0.21.0"
pydantic = "^1.8.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^22.0"
pytest-datadir = "^1.3.1"
pytest-black = "^0.3.12"
pytest-mccabe = "^2.0"
pytest-cov = "^2.11.0"
pytest-mock = "^3.6.1"
sphinx = "^4.0.0"
mypy = "^0.800"
pylint = "^2.7.2"
invoke = "^1.5.0"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
max-args = 6
max-attributes = 15
min-public-methods = 0
disable = "W1201, W1203, R0401, W0707, W0221, W0603, R0801, typecheck"

[build-system]
requires = ["setuptools==59.5.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"